Pivot table for the competition

+++ database/migrations/2021_08_26_100403_create_competition_racers_table.php
@@ -16,7 +16,7 @@ class CreateCompetitionRacersTable extends Migration
             $table->unsignedTinyInteger('place');
             $table->string('car', 50);
             $table->string('result');
-            $table->unsignedTinyInteger('pts');
+            $table->unsignedInteger('pts');
         });
     }
